Biplop

+++ index.js
@@ -120,47 +120,49 @@ async function run() {
       }
     });
 
-// User Profile Update Done By (dev-Arif)
-   app.patch("/userUpdate", async (req, res) => {
-     try {
-       const updatedUser = req.body;
+    // User Profile Update Done By (dev-Arif)
+    app.patch("/user-update", async (req, res) => {
+      try {
+        const updatedUser = req.body;
 
-       if (!updatedUser.email) {
-         return res
-           .status(400)
-           .send({ message: "Email is required for updating a user" });
-       }
+        if (!updatedUser.email) {
+          return res
+            .status(400)
+            .send({ message: "Email is required for updating a user" });
+        }
 
-       const query = { email: updatedUser.email };
-       const existingUser = await userCollection.findOne(query);
+        const query = { email: updatedUser.email };
+        const existingUser = await userCollection.findOne(query);
 
-       if (!existingUser) {
-         return res.status(404).send({ message: "User not found" });
-       }
+        if (!existingUser) {
+          return res.status(404).send({ message: "User not found" });
+        }
 
-       // Remove the email field from the updatedUser to prevent it from being updated
-       delete updatedUser.email;
+        // Remove the email field from the updatedUser to prevent it from being updated
+        delete updatedUser.email;
 
-       const updateResult = await userCollection.updateOne(query, {
-         $set: updatedUser,
-       });
+        const updateResult = await userCollection.updateOne(query, {
+          $set: updatedUser,
+        });
 
-       if (updateResult.modifiedCount === 1) {
-         res.send({ message: "User updated successfully" });
-       } else {
-         res.send({ message: "User not updated" });
-       }
-     } catch (error) {
-       console.log(error);
-       res.status(500).send({ message: "Internal server error" });
-     }
-   });
+        if (updateResult.modifiedCount === 1) {
+          res.send({ message: "User updated successfully" });
+        } else {
+          res.send({ message: "User not updated" });
+        }
+      } catch (error) {
+        console.log(error);
+        res.status(500).send({ message: "Internal server error" });
+      }
+    });
 
     // User Role
     app.get("/user-role/:email", verifyJWT, async (req, res) => {
       try {
         const email = req.params.email;
+        console.log(email);
         const findUser = await userCollection.findOne({ email: email });
+        console.log(findUser);
         res.send({
           userRole: findUser.userRole,
         });
@@ -170,31 +172,44 @@ async function run() {
     });
 
     // Load All Users: (dev-akash)
-    app.get('/user', async (req, res) => {
-      const cursor = userCollection.find();
-      const result = await cursor.toArray();
-      res.send(result);
-    })
+    app.get("/users", async (req, res) => {
+      try {
+        const cursor = userCollection.find();
+        const result = await cursor.toArray();
+        res.send(result);
+      } catch (error) {
+        console.log(error);
+      }
+    });
 
     // Load Single User: (dev-akash)
-    app.get('/user/:id', async (req, res) => {
-      const id = req.params.id;
-      const user = await userCollection.findOne({ _id: new ObjectId(id) });
-      res.send(user);
-    })
+    app.get("/user/:id", async (req, res) => {
+      try {
+        const id = req.params.id;
+        const user = await userCollection.findOne({ _id: new ObjectId(id) });
+        res.send(user);
+      } catch (error) {
+        console.log(error);
+      }
+    });
 
     // User Delete Method: (dev-akash)
-    app.delete('/user/:id', async (req, res) => {
-      const id = req.params.id;
-      const deleteUser = { _id: new ObjectId(id) };
-      const result = await userCollection.deleteOne(deleteUser);
-      if (result.deletedCount) {
-        res.send(result)
-      }
-      else {
-        res.send({ message: 'Something Went Wrong!' })
-      }
-    })
+    app.delete("/user/:id", async (req, res) => {
+      try {
+        const id = req.params.id;
+        const deleteUser = { _id: new ObjectId(id) };
+        if (deleteUser) {
+          const result = await userCollection.deleteOne(deleteUser);
+          if (result.deletedCount) {
+            res.send(result);
+          } else {
+            res.send({ message: "Something Went Wrong!" });
+          }
+        } else {
+          res.send({ message: "Something Went Wrong!" });
+        }
+      } catch (error) {}
+    });
 
     // Job Post: (dev-akash)
     app.post("/job-post", async (req, res) => {
@@ -211,22 +226,32 @@ async function run() {
       }
     });
 
-    // Get All Job Post: (dev-akash)
-    app.get('/job-post', async (req, res) => {
-      const allJobPost = jobsCollection.find();
-      const result = await allJobPost.toArray();
-      res.send(result);
-    })
+    // Get All Job: (dev-akash)
+    app.get("/all-jobs", async (req, res) => {
+      try {
+        const allJobPost = jobsCollection.find();
+        const result = await allJobPost.toArray();
+        res.send(result);
+      } catch (error) {
+        console.log(error);
+      }
+    });
 
     // Load Single Job: (dev-akash)
-    app.get('/job-post/:id', async (req, res) => {
-      const id = req.params.id;
-      const singleJob = await jobsCollection.findOne({ _id: new ObjectId(id) });
-      res.send(singleJob);
+    app.get("/job/:id", async (req, res) => {
+      try {
+        const id = req.params.id;
+        const singleJob = await jobsCollection.findOne({
+          _id: new ObjectId(id),
+        });
+        res.send(singleJob);
+      } catch (error) {
+        console.log(error);
+      }
     });
 
     // Job Update : (dev-Arif)
-    app.patch("/job/:jobId", async (req, res) => {
+    app.patch("/update-job/:jobId", async (req, res) => {
       try {
         const updatedJob = req.body;
         const jobId = req.params.jobId;
@@ -249,18 +274,16 @@ async function run() {
     });
 
     // Job Post Delete Method: (dev-akash)
-    app.delete('/job-post/:id', async (req, res) => {
+    app.delete("/delete-job/:id", async (req, res) => {
       const id = req.params.id;
       const deleteJob = { _id: new ObjectId(id) };
       const result = await jobsCollection.deleteOne(deleteJob);
       if (result.deletedCount) {
-        res.send(result)
+        res.send(result);
+      } else {
+        res.send({ message: "Something Went Wrong!" });
       }
-      else {
-        res.send({ message: 'Something Went Wrong!' })
-      }
-    })
-
+    });
 
     await client.db("admin").command({ ping: 1 });
     console.log(
